Separate the routes from the main server file
